requirements.txt
alembic==1.7.4
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.3
cycler==0.10.0
decorator==5.1.0
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-paginate==2021.10.29
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.3
infinity==1.5
intervals==0.9.2
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.2
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
pycparser==2.20
pyparsing==3.0.3
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.26
validators==0.18.2
Werkzeug==2.0.2
WTForms==2.3.3
WTForms-Components==0.10.5
